Assistant checkpoint: Updated UI with dark mode and suggested questions
Assistant generated file changes:
- templates/index.html: Update UI with new features

---

User prompt:

Add at the bottom my signature CloudCraftWithFranck, put the link to visit my website at https://www.cloudcraftwithfranck.org/. Suggest types of questions as well like in the screenshot. Also put the short paragraph on top: As an engineer, I usually find myself spending a lot of time googling, so I made this tool for IT engineers to help facilitate this process. Add the dark and white mode switch. Also the result of the prompt should be formatted like in the code block.